# LSTM and GRU performance: Replace tf.slice with tf.split