PERF

Improvement in Benchmarks (Averaging 5-6 repeated runs of the benchmark to reduce noise):

- GRU predict:
  - 92.7 ms --> 86.4 ms (1.07x)
- LSTM predict:
  - 72.1 ms --> 68.7 ms (1.05x)
- Attention predict:
  - 72.9 ms --> 69.6 ms (1.05x)

Relating to https://github.com/tensorflow/tfjs/issues/863

<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-layers/365)
<!-- Reviewable:end -->